Qsmtpd: fix crash due to bad truncation
For this to happens some things need to happen at the same time:
-sizeof(long) != sizeof(int) (i.e. 64 bit platform)
-the command has a target state of -1 (i.e. "do not change state")
-this command needs to be accepted and called twice

The target state will be saved and later restored. For this a too short
variable was used, leading to a bad target state afterwards. It was no longer
negative, so after the second command the target state was 0xffffffff, allowing
all commands to happen next.

A sequence that may lead to a crash was:
VRFY
VRFY
MAIL FROM:<>

This happened because MAIL FROM relies on information from HELO or EHLO being
present, which had not happened here.

Found by american fuzzy lop.

svn path=/trunk/; revision=2370

+++ tests/cmd_from_test.c
@@ -54,7 +54,7 @@ string liphost = {
 	.s = (char *)((uintptr_t)-1),
 	.len = -1
 };
-unsigned int comstate = 0x10000000;
+unsigned long comstate = 0x10000000;
 struct smtpcomm *current_command = (struct smtpcomm *)((uintptr_t)-1);
 
 void
